Refactor: Fix bundle type

+++ app/app/src/main/java/com/example/happybankbook/adapter/OnItemSelectedListener.java
@@ -1,5 +1,5 @@
 package com.example.happybankbook.adapter;
 
 public interface OnItemSelectedListener {
-    void onItemSelected(long itemId);
+    void onItemSelected(int itemId);
 }
